Провел рефакторинг composable функций
