+++ app/src/main/java/com/example/traindriver/ui/screen/signin_screen/StartElemants.kt
@@ -1,283 +1,1 @@
 package com.example.traindriver.ui.screen.signin_screen
-
-import android.app.Activity
-import android.util.Log
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
-import com.example.traindriver.R
-import com.example.traindriver.data.auth.SignInMethod
-import com.example.traindriver.data.util.ResultState
-import com.example.traindriver.ui.element_screen.NumberPhoneTextField
-import com.example.traindriver.ui.screen.ScreenEnum
-import com.example.traindriver.ui.theme.ShapeInputData
-import com.example.traindriver.ui.theme.TrainDriverTheme
-import com.example.traindriver.ui.theme.Typography
-import com.example.traindriver.ui.util.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-
-@Composable
-fun StartElements(
-    navController: NavController,
-    signInViewModel: SignInViewModel = viewModel(),
-    activity: Activity
-) {
-    val scope = rememberCoroutineScope()
-    val number = signInViewModel.number
-    val allowEntry = signInViewModel.allowEntry
-    val localeState = signInViewModel.locale
-    val logInToApp = object : LogInToApp {
-        override fun logIn(method: String) {
-            when (method) {
-                SignInMethod.Anonymous.javaClass.name -> {
-                    signInViewModel.signIn()
-                }
-                SignInMethod.Phone.className -> {
-                    scope.launch(Dispatchers.Main) {
-                        signInViewModel.signInWithPhone(activity)
-                            .collect {
-                                when (it) {
-                                    is ResultState.Loading -> {}
-                                    is ResultState.Success -> {
-                                        Log.d("ZZZ", "Открыть экраен ввода СМС Кода")
-                                        // Открыть экраен ввода СМС Кода
-                                    }
-                                    is ResultState.Failure -> {}
-                                }
-                            }
-                    }
-                }
-                SignInMethod.Google.javaClass.name -> {}
-                SignInMethod.Email.javaClass.name -> {}
-            }
-
-        }
-    }
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.SpaceAround
-    ) {
-        Logo()
-        InputDataElements(number, allowEntry, localeState, logInToApp)
-        SkipButton(navController = navController, logInToApp = logInToApp)
-    }
-}
-
-@Composable
-fun Logo() {
-    Text(
-        modifier = Modifier.fillMaxWidth(),
-        text = "Машинистам",
-        style = Typography.h1,
-        color = MaterialTheme.colors.onPrimary
-    )
-}
-
-@Composable
-fun InputDataElements(
-    number: MutableState<String>,
-    allowEntry: MutableState<Boolean>,
-    localeState: MutableState<LocaleUser>,
-    logInToApp: LogInToApp
-) {
-    Surface(
-        modifier = Modifier.wrapContentSize(),
-        shape = ShapeInputData.medium,
-        color = MaterialTheme.colors.surface,
-        elevation = dimensionResource(id = R.dimen.elevation_input_element)
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(dimensionResource(id = R.dimen.padding_all_input_element))
-                .fillMaxWidth(0.85f),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = stringResource(id = R.string.title_input_element),
-                style = Typography.h3,
-                color = MaterialTheme.colors.onBackground
-            )
-
-            PrimarySpacer()
-            NumberPhoneTextField(
-                placeholderText = stringResource(id = R.string.placeholder_input_number),
-                numberState = number,
-                localeUser = localeState,
-                allowEntry = allowEntry,
-                isFilledCallback = if (localeState.value != LocaleUser.OTHER) {
-                    object : FieldIsFilled {
-                        override fun isFilled(isFilled: Boolean) {
-                            allowEntry.value = isFilled
-                        }
-                    }
-                } else null
-            )
-
-            SecondarySpacer()
-            LoginButton(
-                enabled = allowEntry.value,
-                logInToApp = logInToApp
-            )
-
-            PrimarySpacer()
-            DividerTrainDriver()
-
-            PrimarySpacer()
-            AlternativeEnteringMenu(logInToApp)
-        }
-    }
-}
-
-@Composable
-fun AlternativeEnteringMenu(logInToApp: LogInToApp) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        CircleButton(
-            R.drawable.ic_launcher_foreground, stringResource(id = R.string.cont_desc_google_button)
-        ) {
-            logInToApp.logIn(SignInMethod.Google.javaClass.name)
-            Log.d("Debug", "click button login with Google")
-        }
-        CircleButton(
-            R.drawable.ic_launcher_foreground, stringResource(id = R.string.cont_desc_email_button)
-        ) {
-            logInToApp.logIn(SignInMethod.Email.javaClass.name)
-            Log.d("Debug", "click button login by Email")
-        }
-        CircleButton(
-            R.drawable.ic_launcher_foreground, stringResource(id = R.string.cont_desc_vk_button)
-        ) {
-            Log.d("Debug", "click button login by Email")
-        }
-    }
-}
-
-@Composable
-private fun DividerTrainDriver() {
-    Row(
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Divider(
-            modifier = Modifier
-                .weight(1f)
-                .padding(end = dimensionResource(id = R.dimen.padding_divider)),
-            color = MaterialTheme.colors.onSurface
-        )
-        Text(
-            modifier = Modifier.wrapContentWidth(),
-            textAlign = TextAlign.Justify,
-            text = stringResource(id = R.string.divider_text),
-            style = Typography.body2,
-            color = MaterialTheme.colors.onSurface
-        )
-        Divider(
-            modifier = Modifier
-                .weight(1f)
-                .padding(start = dimensionResource(id = R.dimen.padding_divider)),
-            color = MaterialTheme.colors.onSurface
-        )
-
-    }
-}
-
-@Composable
-fun PrimarySpacer() {
-    Spacer(
-        modifier = Modifier.padding(dimensionResource(id = R.dimen.primary_padding_between_view))
-    )
-}
-
-@Composable
-fun SecondarySpacer() {
-    Spacer(
-        modifier = Modifier.padding(dimensionResource(id = R.dimen.secondary_padding_between_view))
-    )
-}
-
-@Composable
-fun LoginButton(
-    enabled: Boolean,
-    logInToApp: LogInToApp
-) {
-    Button(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(dimensionResource(id = R.dimen.min_size_view)),
-        onClick = {
-            logInToApp.logIn(SignInMethod.Phone.className)
-        },
-        colors = ButtonDefaults.buttonColors(
-            backgroundColor = MaterialTheme.colors.secondaryVariant,
-            disabledBackgroundColor = MaterialTheme.colors.secondary
-        ),
-        enabled = enabled
-    ) {
-        Text(
-            text = stringResource(id = R.string.text_entrance_button),
-            color = MaterialTheme.colors.onSecondary,
-            style = Typography.button
-        )
-    }
-}
-
-@Composable
-fun CircleButton(resId: Int, contentDescription: String?, action: () -> Unit) {
-    Button(
-        modifier = Modifier.size(dimensionResource(id = R.dimen.min_size_view)),
-        shape = CircleShape,
-        colors = ButtonDefaults.buttonColors(backgroundColor = Color.White),
-        onClick = action
-    ) {
-        Image(
-            painter = painterResource(id = resId),
-            contentDescription = contentDescription,
-            contentScale = ContentScale.Crop
-        )
-    }
-}
-
-@Composable
-fun SkipButton(navController: NavController, logInToApp: LogInToApp) {
-    Text(
-        modifier = Modifier.clickable(onClick = {
-            logInToApp.logIn(method = SignInMethod.Anonymous.javaClass.name)
-            navController.apply {
-                this.popBackStack(ScreenEnum.SIGN_IN.name, true)
-                this.navigate(ScreenEnum.MAIN.name)
-            }
-        }),
-        text = stringResource(id = R.string.text_skip_button),
-        style = Typography.body2,
-        color = MaterialTheme.colors.onBackground
-    )
-}
-
-@Composable
-@DarkLightPreviews
-@FontScalePreviews
-private fun StartScreenPrev() {
-    TrainDriverTheme {
-//        StartElements(localeState = mutableStateOf(LocaleState.RU))
-    }
-}
