+++ app/src/main/java/com/example/traindriver/ui/screen/signin_screen/elements/LoginButton.kt
@@ -1,0 +1,50 @@
+package com.example.traindriver.ui.screen.signin_screen.elements
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.stringResource
+import com.example.traindriver.R
+import com.example.traindriver.ui.theme.TrainDriverTheme
+import com.example.traindriver.ui.theme.Typography
+import com.example.traindriver.ui.util.DarkLightPreviews
+import com.example.traindriver.ui.util.FontScalePreviews
+
+@Composable
+fun LoginButton(
+    enabled: Boolean,
+    onClick: () -> Unit
+) {
+    Button(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(dimensionResource(id = R.dimen.min_size_view)),
+        onClick = onClick,
+        colors = ButtonDefaults.buttonColors(
+            backgroundColor = MaterialTheme.colors.secondaryVariant,
+            disabledBackgroundColor = MaterialTheme.colors.secondary
+        ),
+        enabled = enabled
+    ) {
+        Text(
+            text = stringResource(id = R.string.text_entrance_button),
+            color = MaterialTheme.colors.onSecondary,
+            style = Typography.button
+        )
+    }
+}
+
+@Composable
+@DarkLightPreviews
+@FontScalePreviews
+private fun StartScreenPrev() {
+    TrainDriverTheme {
+        LoginButton(enabled = true, {})
+    }
+}
