+++ app/src/main/java/com/example/traindriver/ui/screen/signin_screen/elements/Logo.kt
@@ -1,0 +1,30 @@
+package com.example.traindriver.ui.screen.signin_screen.elements
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.example.traindriver.ui.theme.TrainDriverTheme
+import com.example.traindriver.ui.theme.Typography
+import com.example.traindriver.ui.util.DarkLightPreviews
+import com.example.traindriver.ui.util.FontScalePreviews
+
+@Composable
+fun Logo(modifier: Modifier) {
+    Text(
+        modifier = modifier,
+        text = "Машинистам",
+        style = Typography.h1,
+        color = MaterialTheme.colors.onPrimary
+    )
+}
+
+@Composable
+@DarkLightPreviews
+@FontScalePreviews
+private fun StartScreenPrev() {
+    TrainDriverTheme {
+        Logo(modifier = Modifier.fillMaxWidth())
+    }
+}
