--- app/src/main/java/com/example/traindriver/ui/util/LogInToApp.kt
@@ -1,5 +1,0 @@
-package com.example.traindriver.ui.util
-
-interface LogInToApp {
-    fun logIn(method: String)
-}
